+++ application/shared/view/404.php
@@ -1,0 +1,11 @@
+<html>
+    <head>
+        <title>404 Error</title>
+    </head>
+    <body>
+        <div style="width: 30%; margin: 0 auto; text-align: center;">
+            <h1>404 Error</h1>
+            <h2>File Not Found</h2>
+        </div>
+    </body>
+</html>